Generate GH-pages HTML files again
This reverts commit d6055e3cd89d05cf0c87ebeeb590178d087267c8
and partially reverts commit
807eebb019cf79ed129b785c80105dd6ecbe60f2.

The HTML files that are expected to be deployed to GitHub Pages are
now located in the `static` directory, and copied to dist/ on
build to keep the URLs the same.

Note that these static files are published to npm as well. For more
background on this, see
https://github.com/solid/mashlib/pull/61#issuecomment-522943381